// iuffimobile/src/app/detections/dto/plant-species-aviv.dto.ts
export interface PlantSpeciesAvivDto {
	denomAutorizzazione: string;
	denomAttivita: string;
	specieNomeBotanico: string;
	specieNomeVolgare: string;
	flgPassUpp: boolean;
	flgPassZp: boolean;
	flgAccr: boolean;
	flgColtCampo: boolean;
	flgColtCont: boolean;
	flgColtSerra: boolean;
	qtaProduzione: number;
	unitMisura: string;
}